added last name generator
JS - added a last name generator that creates a last name based on
three separate arays(similar to the first name).

moved the appendCharacter() into the npcCreate to reduce confusion when
the program refers to the character.

CSS - changed the sizing of the form to match new sizing of the
npcCharacterInfoWrapper.